pytorch-deployment-tutorial/requirements.txt
numpy
Pillow
torch>=1.5.0
torchvision>=0.6.0